feat: get example working again
There's still a ton of CPU usage, and it's not as fast as I'd like it to be. I
should look into leveraging webassembly? Or at least getting rid of some of the
garbage collection - etc...

I should also look into setting up unit tests, but for the life of me I have no
idea how they would work. Basically, I'd need integration tests using real ROMs?
I'm unlikely to get to do that any time soon.

For now, this will work. Continue cleanup, package it up just a bit, and let it
go from there.

The whole point is that this will run on a separate device and you'll connect
over socket to control it. So it's OK to handle everything that way. You should
have a socket request to get the next frame or something.